refactor(ledger): serve dust_generation_values from the ledger-8 host api
The read only ever touches ledger-8 state, so move it off Ledger9Bridge onto
Ledger8Bridge, drop the now-redundant _v8 suffix, and move its return types
out of common::types into host_api::ledger_8.
